[FIX] correction filecontent ServiceAddJob.cd ServiceLIstSaveJob.cs for merge
