+++ CLI/src/CommandeListJobs.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Services;
 
 namespace EasySave
 {
@@ -8,9 +9,7 @@ namespace EasySave
 
         public override void Action(string[] args)
         {
-            Configuration configuration = new Configuration( Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData) + "\\EsaySave\\" + "config.json");
-            configuration.LoadConfiguration();
-            SaveJob[] saveJobs = configuration.GetSaveJobs();
+            SaveJob[] saveJobs = ServiceListSaveJob.Run(args);
             foreach (var saveJob in saveJobs)
             {
                 Console.WriteLine(ConsoleColors.Bold + ConsoleColors.Yellow + "\tName : " + saveJob.Name + ConsoleColors.Reset);
