+++ Services/Services.csproj
@@ -6,4 +6,8 @@
         <Nullable>enable</Nullable>
     </PropertyGroup>
 
+    <ItemGroup>
+      <ProjectReference Include="..\Config\Config.csproj" />
+    </ItemGroup>
+
 </Project>
